+++ src/Paynow/Util/SignatureCalculator.php
@@ -13,10 +13,10 @@ class SignatureCalculator
 
     /**
      * @param string $signatureKey
-     * @param array  $data
+     * @param string  $data
      * @throws InvalidArgumentException
      */
-    public function __construct($signatureKey, array $data)
+    public function __construct($signatureKey, $data)
     {
         if (empty($signatureKey)) {
             throw new InvalidArgumentException('You did not provide a Signature key');
@@ -25,7 +25,7 @@ class SignatureCalculator
         if (empty($data)) {
             throw new InvalidArgumentException('You did not provide any data');
         }
-        $this->hash = base64_encode(hash_hmac('sha256', json_encode($data, JSON_UNESCAPED_SLASHES), $signatureKey, true));
+        $this->hash = base64_encode(hash_hmac('sha256', $data, $signatureKey, true));
     }
 
     /**